restructure file tree to conform to convention
