--- src/components/Contact.js
@@ -1,79 +1,0 @@
-import emailjs from 'emailjs-com';
-import React from 'react';
-
-export default function ContactUs() {
-  function sendEmail(e) {
-    e.preventDefault();
-
-    emailjs
-      .sendForm(
-        'service_1nwpeyk',
-        'template_g9g42q9',
-        e.target,
-        'user_P07Tx1xVAjmyvUS92nW0A'
-      )
-      .then(
-        (result) => {
-          console.log(result.text);
-        },
-        (error) => {
-          console.log(error.text);
-        }
-      );
-    e.target.reset();
-  }
-
-  return (
-    <>
-      <div>
-        <div className='container'>
-          <form onSubmit={sendEmail}>
-            <div className='row pt-5 mx-auto'>
-              <div className='col-8 form-group mx-auto'>
-                <input
-                  type='text'
-                  className='form-control'
-                  placeholder='Name'
-                  name='name'
-                />
-              </div>
-              <div className='col-8 form-group pt-2 mx-auto'>
-                <input
-                  type='email'
-                  className='form-control'
-                  placeholder='Email Address'
-                  name='email'
-                />
-              </div>
-              <div className='col-8 form-group pt-2 mx-auto'>
-                <input
-                  type='text'
-                  className='form-control'
-                  placeholder='Subject'
-                  name='subject'
-                />
-              </div>
-              <div className='col-8 form-group pt-2 mx-auto'>
-                <textarea
-                  className='form-control'
-                  id=''
-                  cols='30'
-                  rows='8'
-                  placeholder='Your message'
-                  name='message'
-                ></textarea>
-              </div>
-              <div className='col-8 pt-3 mx-auto'>
-                <input
-                  type='submit'
-                  className='btn btn-info'
-                  value='Send Message'
-                ></input>
-              </div>
-            </div>
-          </form>
-        </div>
-      </div>
-    </>
-  );
-}
